Prepare for publish
Move examples into a separate folder.
Update index to expose the functionality.
