+++ examples/semaphore.ts
